+++ src/PartialProperties.ts
@@ -1,0 +1,3 @@
+export type PartialProperties<T> = {
+  [F in keyof T]?: T[F];
+};